// Public/nextjs/Public/src/models/VariantType.Model.ts
export interface IVariantTypeModel {
    Active: boolean
    Created: string
    CreatedBy: number
    DateOfIntroduction: any
    DisplayInList: boolean
    DisplayName: string
    EnumName: string
    Id: number
    IsValid: boolean
    MaximumValue: number
    MinimumValue: number
    Modified: string
    ModifiedBy: number
    OrderWishlistId: number
    Name: string
    RowOrder: number
    SeriesModelId: number
    ThumbnailPath: any
    TotalRowsCount: any
    ValidationErrors: any
}